+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Remove **page** option
     - Add **output** option to generate PDF [#6]
 - CSS correction [#13]
+- Add **Table of Content** in the CSS Squeleton
 
 ## [0.3.0] - 2021-08-08
 ### Added